remove editor fie
