--- src/Features/Core/Portable/GenerateMember/GenerateConstructor/AbstractGenerateConstructorService.Editor.cs
@@ -1,54 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Collections.Generic;
-using System.Collections.Immutable;
-using System.Diagnostics;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.CodeGeneration;
-using Microsoft.CodeAnalysis.Editing;
-using Microsoft.CodeAnalysis.Shared.Extensions;
-
-namespace Microsoft.CodeAnalysis.GenerateMember.GenerateConstructor
-{
-    internal abstract partial class AbstractGenerateConstructorService<TService, TArgumentSyntax, TAttributeArgumentSyntax>
-    {
-        protected abstract bool IsConversionImplicit(Compilation compilation, ITypeSymbol sourceType, ITypeSymbol targetType);
-
-        internal abstract IMethodSymbol GetDelegatingConstructor(State state, SemanticDocument document, int argumentCount, INamedTypeSymbol namedType, ISet<IMethodSymbol> candidates, CancellationToken cancellationToken);
-
-        protected abstract IMethodSymbol GetCurrentConstructor(SemanticModel semanticModel, SyntaxToken token, CancellationToken cancellationToken);
-
-        protected abstract IMethodSymbol GetDelegatedConstructor(SemanticModel semanticModel, IMethodSymbol constructor, CancellationToken cancellationToken);
-
-        protected bool CanDelegeteThisConstructor(State state, SemanticDocument document, IMethodSymbol delegatedConstructor, CancellationToken cancellationToken = default)
-        {
-            var currentConstructor = GetCurrentConstructor(document.SemanticModel, state.Token, cancellationToken);
-            if (currentConstructor.Equals(delegatedConstructor))
-            {
-                return false;
-            }
-
-            // We need ensure that delegating constructor won't cause circular dependency.
-            // The chain of dependency can not exceed the number for constructors
-            var constructorsCount = delegatedConstructor.ContainingType.InstanceConstructors.Length;
-            for (var i = 0; i < constructorsCount; i++)
-            {
-                delegatedConstructor = GetDelegatedConstructor(document.SemanticModel, delegatedConstructor, cancellationToken);
-                if (delegatedConstructor == null)
-                {
-                    return true;
-                }
-
-                if (delegatedConstructor.Equals(currentConstructor))
-                {
-                    return false;
-                }
-            }
-
-            return false;
-        }
-    }
-}
